# gateway/src/main/java/com/undina/gateway/entity/LoginRequest.java
//package com.undina.gateway.entity;
//
//import lombok.Data;
//
//import javax.validation.constraints.NotBlank;
//
//@Data
//public class LoginRequest {
//
//    @NotBlank
//    private String email;
//
//    @NotBlank
//    private String password;
//}